created api

+++ backend/src/app/controllers/CourseController.js
@@ -1,4 +1,5 @@
 const Course = require("../models/Course");
+const Lesson = require("../models/Lesson");
 const {
   multipleMongooseToObject,
   mongooseToObject,
@@ -37,15 +38,15 @@ class CourseController {
   async store(req, res, next) {
     try {
       const course = new Course(req.body);
-      // const user = req.user; // Lấy user từ request (sau khi đã authenticate)
+      const user = req.user; // Lấy user từ request (sau khi đã authenticate)
 
-      // // Kiểm tra nếu user không phải admin hoặc instructor thì từ chối
-      // if (user.role !== "admin" && user.role !== "instructor") {
-      //   return res.status(403).json({
-      //     success: false,
-      //     message: "Only admin and instructor can create courses",
-      //   });
-      // }
+      // Kiểm tra nếu user không phải admin hoặc instructor thì từ chối
+      if (user.role !== "admin" && user.role !== "instructor") {
+        return res.status(403).json({
+          success: false,
+          message: "Only admin and instructor can create courses",
+        });
+      }
 
       await course.save();
       res.status(201).json({
@@ -59,15 +60,15 @@ class CourseController {
 
   async destroy(req, res, next) {
     try {
-      // const user = req.user; // Lấy user từ request (sau khi đã authenticate)
+      const user = req.user; // Lấy user từ request (sau khi đã authenticate)
 
-      // // Kiểm tra nếu user không phải admin hoặc instructor thì từ chối
-      // if (user.role !== "admin") {
-      //   return res.status(403).json({
-      //     success: false,
-      //     message: "Only admin can create courses",
-      //   });
-      // }
+      // Kiểm tra nếu user không phải admin hoặc instructor thì từ chối
+      if (user.role !== "admin") {
+        return res.status(403).json({
+          success: false,
+          message: "Only admin can create courses",
+        });
+      }
 
       const course = await Course.findByIdAndDelete({ _id: req.params.id });
       if (!course) {
@@ -86,15 +87,15 @@ class CourseController {
 
   async update(req, res, next) {
     try {
-      // const user = req.user; // Lấy user từ request (sau khi đã authenticate)
+      const user = req.user; // Lấy user từ request (sau khi đã authenticate)
 
-      // // Kiểm tra nếu user không phải admin hoặc instructor thì từ chối
-      // if (user.role !== "admin" && user.role !== "instructor") {
-      //   return res.status(403).json({
-      //     success: false,
-      //     message: "Only admin and instructor can create courses",
-      //   });
-      // }
+      // Kiểm tra nếu user không phải admin hoặc instructor thì từ chối
+      if (user.role !== "admin" && user.role !== "instructor") {
+        return res.status(403).json({
+          success: false,
+          message: "Only admin and instructor can create courses",
+        });
+      }
 
       const courseId = req.params.id;
       const courseData = req.body;
@@ -116,6 +117,25 @@ class CourseController {
       next(err);
     }
   }
+
+  async getLessonsByCourse(req, res, next) {
+    try {
+      const courseID = req.params.id; // Lấy courseId từ params
+      const lessons = await Lesson.find({course_id: courseID});
+      if (!lessons) {
+        return res
+          .status(404)
+          .json({ success: false, message: "Lessons not found" });
+      }
+      res.json({
+        success: true,
+        data: multipleMongooseToObject(lessons),
+      });
+    } catch (err) {
+      next(err);
+    }
+  }
+
 }
 
 module.exports = new CourseController();
